# Improvement: Remove Cached JSON File Reading for Faster Performance