@solomon-ochepa 2024-06-24
Regarding the issue https://github.com/nWidart/laravel-modules/issues/1745, completely eradicating the caching system is not advisable. The ability to disable or enable it when needed is sufficient. By default, it is disabled, which effectively makes it inactive, but it allows those who find it useful to activate it when necessary.

I recommend testing other scenarios, such as those mentioned in https://github.com/nWidart/laravel-modules/issues/1745#issuecomment-2122390719, https://github.com/nWidart/laravel-modules/issues/1745#issuecomment-2021464561, https://github.com/nWidart/laravel-modules/issues/1745#issuecomment-1987761705, https://github.com/nWidart/laravel-modules/issues/1745#issuecomment-1965250076, for potential improvements.

When proposing a change or feature like this, consider its impact on others. The best practice is to provide a configuration option to enable or disable the feature as needed, and this element meets that requirement.

Thank you for your contribution.