@alissn 2024-06-24
@solomon-ochepa

Did you review the changes and test this pull request?

This pull request only deletes the cache of the JSON file of the modules and does not address the mentioned issue.