@solomon-ochepa 2024-06-24
> @solomon-ochepa
> 
> Did you review the changes and test this pull request?
> 
Yes, sir. I did.

> This pull request only deletes the cache of the JSON file of the modules (NOT ALL CACHE) and does not address the mentioned issue.

![image](https://github.com/nWidart/laravel-modules/assets/110875083/d9544095-07ca-40ec-afc1-f5bb94069911)

You asked that the `getAttributes()` method return `$this->decodeContents()` without checking the condition - `if (config('modules.cache.enabled') === false)`.

"The ability to disable or enable it when needed is sufficient. By default, it is disabled, which effectively makes it inactive, but it allows those who find it useful to activate it when necessary."