@alissn 2024-06-24
@solomon-ochepa

Please follow the "How to Test This" section in the description and share your results.

The default cache driver is `file`, which creates a cache file from the JSON file and reads from the cached file using the cache driver. Do you think this is correct and should be kept?