@solomon-ochepa 2024-06-24
> @solomon-ochepa
> 
> Please follow the "How to Test This" section in the description and share your results.
> 
> The default cache driver is `file`, which creates a cache file from the JSON file and reads from the cached file using the cache driver. Do you think this is correct and should be kept?

Your test ignored the condition `if(config("modules.cache.enabled" == false){...}` and read from cache directly -  that is wrong testing!

"... By default, it is disabled, which effectively makes it inactive, but it allows those who find it useful to activate it when necessary."