@dcblogdev 2024-06-29
Reading from file is constantly faster, I ran 150 modules and ran the benchmarks

```
array:2 [▼
  "directly from file" => "0.013ms"
  "read from cache" => "0.166ms"
]

array:2 [▼
  "directly from file" => "0.013ms"
  "read from cache" => "0.078ms"
]

array:2 [▼
  "directly from file" => "0.013ms"
  "read from cache" => "0.077ms"
]

array:2 [▼
  "directly from file" => "0.014ms"
  "read from cache" => "0.063ms"
]

array:2 [▼
  "directly from file" => "0.013ms"
  "read from cache" => "0.078ms"
]

array:2 [▼
  "directly from file" => "0.013ms"
  "read from cache" => "0.080ms"
]
```